MyProject.Test/API.Test/Controller.Test/UserGroupControllerTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Moq;
using MyProject.Api.Controllers;
using MyProject.Application.Interfaces;
using MyProject.Application.IServices;
using MyProject.Domain.Entities;
using Xunit;

namespace MyProject.Tests.Controllers
{
    public class UserGroupControllerTests
    {
        [Fact]
        public async Task GetAllUserGroups_ReturnsOkResult()
        {
            // Arrange
            var mockService = new Mock<IUserGroupService>();
            mockService.Setup(service => service.GetAllAsync()).ReturnsAsync(new List<UserGroup>());

            var controller = new UserGroupController(mockService.Object);

            // Act
            var actionResult = await controller.GetAllUserGroups();
            var result = actionResult as OkObjectResult; // Checking if the result is an OkObjectResult
                                                         // Assert
            Assert.NotNull(result);
        }

        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetUserGroupById_ExistingId_ReturnsOkResult(int userGroupId)
        {
            // Arrange
            var mockService = new Mock<IUserGroupService>();
            mockService.Setup(service => service.GetByIdAsync(userGroupId)).ReturnsAsync(new UserGroup { Id = userGroupId, Name = "TestGroup" });

            var controller = new UserGroupController(mockService.Object);

            // Act
            var result = await controller.GetUserGroupById(userGroupId);

            // Assert
            Assert.NotNull(result);
        }

        [Fact]
        public async Task CreateUserGroup_ValidObject_ReturnsCreatedAtAction()
        {
            // Arrange
            var mockService = new Mock<IUserGroupService>();
            var newUserGroup = new UserGroup { Name = "New Group" }; // Replace with test data
            mockService.Setup(service => service.CreateAsync(newUserGroup)).ReturnsAsync(1); // Replace with an expected ID

            var controller = new UserGroupController(mockService.Object);

            // Act
            var result = await controller.CreateUserGroup(newUserGroup) as CreatedAtActionResult;

            // Assert
            Assert.NotNull(result);
            Assert.Equal(nameof(UserGroupController.GetUserGroupById), result.ActionName);
            Assert.Equal(1, result.RouteValues["id"]); // Replace with expected route value
        }

        // Write similar tests for UpdateUserGroup and DeleteUserGroup actions

        // Remember to add specific assertions to validate the responses and behaviors of each action.
    }
}
